requirements.txt
flask
flask-socketio
python-socketio[client]
requests
python-dotenv
openai